work on bcgov/name-examination#736 bcgov/name-examination#737
There was a change in the keycloak map, username -> preferred_username.
Updating the User.create_from_jwtToken method to use the 'preferred_username'

update PUT to use the same calls for security and segments as Patch
not in this go-round, but the goal is that Put & Patch are almost identical, with marshmallow set to partial updates only in Patch, and maybe difference's in edit approval.