bcgov/name-examination#739
In the distributed system there are times that some of the remote services will fail, but the business flow can continue. In those cases we still return or process the object, but provide the back-office users warning messages. To make it easier for the clients to process, we will change that return from a 200 to a 206 to indicate that there are warnings that may require alternate business processes.

Change return of Requests-PUT/PATCH when a warning exists to a 206